Fixing test to match existing behavior
The test previously failed when http was fromGlobals() because the Uri
uses a default host if it's set.  Since it is proper to set 'http' by
default in `fromGlobals()`, I left the existing default host behavior
as is.  A Request or ServerRequest even not populated from globals
would not necessarily suffer the same fate.