Add Header::splitList()
Header::splitList() is a replacement for the existing normalize() method with a
cleaned up API, proper explanation and a more specific method name instead of
the vague `normalize()` that could refer to anything.

The primary difference in the API is that normalize() accepts `string[][]`,
whereas `splitList()` is expected to take a `string[]`: The return value of a
call to `MessageInterface::getHeader()`. Alternatively a `string` may be
passed, corresponding to the return value of
`MessageInterface::getHeaderLine()`.

Further `splitList()` will properly verify the types and throw a `TypeError` if
the input is not well-defined. This allows for a clean migration to a
`string|array` parameter type if the minimum version of guzzle/psr7 is
increased to PHP 8.