Fix deprecation warnings in PHP 8.2
Declare explicit properties where possible:

Any class that is final and sets the `$stream` property in the constructor can
have the property defined, because it is impossible for there to be child
classes that rely on an overriden constructor that leverages the lazy
initialization of the stream.

I opted to make this property private which is a small BC break, because
implicitly created properties are effectively public.

For `LazyOpenStream` and `FnStream` I've opted to use the
`\AllowDynamicProperties` attribute to minimize the amount of changed required.
This should be cleaned up in the next major.

Resolves guzzle/psr7#482
Closes guzzle/psr7#518
see guzzle/psr7#516