Replace `sha1(uniqid())` by `bin2hex(random_bytes(20))`
The output size and format is identical, but `random_bytes()` provides much
better randomness than `uniqid()` even with the $more_entropy parameter which
adds only 32 Bits of entropy on top of the current (micro)time which is more or
less known externally.

For MultipartStream that might result in a guessable boundary, which might
allow an attacker to modify the semantics of an outgoing request with a
well-crafted file.